feat(core):Enhanced API class and removed PHP references
The API class within the wolf_core module has been updated to include more detailed member variables and additional documentation related to their use. The main purpose of this update is to improve usability of the API class.

Also, removed irrelevant PHP comment from README as the core module is written in Python not PHP.