feat(api): Add optional params to GET requests, and add PUT, DELETE methods
Extended method for GET requests in the API to accept optional parameters.
Further added new methods to support PUT and DELETE requests to the API.
Implemented a RequestResponse class to store response data from API request, including status code and data. These changes provide a complete set of methods for interacting with the API, resulting in a flexible and robust API handler.