Upgrade version and add workflows for Test PyPI publish
Increased the version number of the Wolf Core package and added a GitHub action workflow to automatically publish the package to Test PyPI whenever a new release is created. A more detailed README was implemented as well with project descriptions, installation guidelines, documentation and contribution instructions. The changes were made to automate package distribution and provide clearer instructions to potential contributors.