Update README documentation links and fix typo
The README.md file was updated to change the documentation links from 'wolf-eirlab-community' to 'wolf-project'. This change has been implemented to reflect the updated project documentation site.

Moreover, a typo in the French description of the project has also been fixed. Specifically, 'associatins' was corrected to 'associations'. This was done to ensure the accuracy and professionalism of the project description.