Simplify installation process and enhance README docs
Removed developer-specific installation instructions in README files for 'wolf' and 'wolf-core' projects to simplify and make it more user-friendly. This benefits both users and new project contributors by making the installation process more straightforward.

Updated '[packages]' in pyproject.toml from develop = true to develop = false, which suggests adjusting the settings for a production environment.

Lastly, made slight revisions to the online documentation URLs, which may be due to a reorganization of the documentation site.