Update version in pyproject.toml and modify publish.yaml workflow
Updated the version number in pyproject.toml from 1.1 to 2.0, marking a new major release. The dependency 'cffi' was removed from the project as it is no longer needed.

Modifications were made to the publish.yaml workflow. A workflow_dispatch event was added, allowing the workflow to be manually triggered from the GitHub Actions UI, providing more flexibility in controlling when the workflow should run.