Remove use of ReadtheDocs and related files
This commit removes the integration with ReadtheDocs, wiping out the documentation files focused on generating and maintaining the documentation in ReadtheDocs. It reflects a decision to change our documentation strategy, moving away from ReadtheDocs to lean on another documentation platform which better suits our needs. The benefits of this include more control over our documentation process and style, and easier maintainability.